Crate hyperon_atom

Contains types and functions which are necessary to create and manipulate separate atoms. There are four kinds or meta-types of atoms: Atom::Symbol, Atom::Variable, Atom::Grounded and Atom::Expression.

Atom::Symbol represents a single concepts which is identified by name. Two symbols which have the same name reference the same concept.

Atom::Variable represents a variable like a variable in the pattern. Variable itself doesn’t keep its value inside it but instead can be used as a key in a matcher::Bindings to get the real value.

Atom::Grounded represents sub-symbolic knowledge. On the API level it allows keeping data and behaviour inside an atom. There are three aspects of the grounded atom which can be customized:

the type of grounded atom is provided by the atom itself;

matching algorithm of the atom can be modified by the user;

atom can be made executable; such atom can be used to apply some sub-symbolic operations to other atoms as arguments.

Atom::Expression combines other kinds of atoms including expressions themselves.

All kinds of atoms except grounded can be relatively simply created using corresponding constructors: Atom::sym, Atom::var, Atom::expr. Grounded atoms are more complex because they can represent any Rust value.

There are two flavors of Grounded atoms. When Rust value implements AutoGroundedType trait (which is automatically implemented for any value implementing 'static + PartialEq + Clone + Debug) then one can easily wrap it into atom using Atom::value constructor. Resulting atom contains the value, has autogenerated type Atom::sym(<name of T>) where T is a Rust type of the value. This atom is not executable and matches with other grounded atom when PartialEq::eq returns true for the values inside these atoms.

To override any aspect of this predefined grounded atom behaviour, one should make CustomGroundedType implementable for the Rust value. In other words one should implement two additional traits: Grounded and Display. Display is used to represent grounded atom as a string. By implementing Grounded one can make atom executable, implement custom matching logic or return a custom type of the atom. See example in Grounded section.

There are predefined functions which can be used to reuse default behaviour if needed:

rust_type_atom - return Rust type name calculated by compiler;

match_by_equality - match two atoms when PartialEq::eq returns true;

Modules

gnd

matcher

Module contains functions to match atoms and work with variable bindings.

serial

subexpr

Contains algorithms to walk through subexpressions of expression in different ways.

Macros

bind

Constructs new instance of the Bindings with predefined content. Macros takes variable/value pairs as arguments. If value is a single variable then the pair means variable equality. Otherwise pair means assigning value. May be ineffective, should be used mainly in unit tests.

bind_set

Constructs a new BindingsSet with predefined content. Macros takes variable/value pairs as arguments. If value is a single variable then the pair means variable equality. Otherwise pair means assigning value. May be ineffective, should be used mainly in unit tests.

expr

Constructs new Atom using symplified syntax for expressions. Macros has a perfomance penalty because it creates and uses an additional wrapper for grounded atoms. It is intended for using mainly in the unit tests. Use string literals for symbols, identifiers for variables, braces for grounded symbols and S-expressions syntax for expressions.

sym

Constructs new symbol atom. Can be used to construct const instances.

Structs

AtomIter

Iterator of the sub-atoms of the Atom.

AtomIterMut

Mutable iterator of the sub-atoms of the Atom.

ExpressionAtom

An expression atom structure.

SymbolAtom

A symbol atom structure.

VariableAtom

A variable atom structure

Enums

Atom

Atoms are main components of the atomspace. There are four meta-types of atoms: symbol, expression, variable and grounded.

ExecError

Grounded function execution error.

Traits

AutoGroundedType

Alias for the list of traits required for the standard Rust types to be automatically wrapped into GroundedAtom. It is implemented automatically when type implements 'static + PartialEq + Clone + Debug. No need to implement its manually.

ConvertingSerializer

Converts serializable grounded atom into native Rust type T. Returning error from this serializer is not expected as usually it should just write data into a memory buffer for conversion. Default is required to instantiate it in ConvertingSerializer::convert method.

CustomExecute

Trait for implementing custom execution logic. Using this trait one can represent a grounded function as an atom. In order to make it work one should also implement Grounded::as_execute method.

CustomGroundedType

Alias for the list of traits required for a custom Rust grounded type to be successfully wrapped into GroundedAtom. It is implemented automatically when type implements AutoGroundedType + Display + Grounded. No need to implement it manually. Implement Grounded and Display instead.

CustomMatch

Trait for implementing custom matching logic. In order to make it work one should also implement Grounded::as_match method.

Grounded

Trait allows implementing grounded atom with custom behavior. If no custom behavior is needed then simpler way is using Atom::value function for automatic grounding.

GroundedAtom

A trait to erase an actual type of the grounded atom. Not intended to be implemented by users. Use Atom::value or implement Grounded and use Atom::gnd instead.

Functions

make_variables_unique

Returns atom with all variables replaced by unique instances.

match_by_equality

Returns either single empty matcher::Bindings instance if self and other are equal or empty iterator if not. This is a default implementation of match_() for the grounded types wrapped automatically.

match_by_string_equality

Returns either single empty matcher::Bindings instance if the string representing self and other render are identical strings, or an empty iterator if not.

rust_type_atom

Returns the name of the Rust type wrapped into Atom::Symbol. This is a default implementation of type_() for the grounded types wrapped automatically.

Type Aliases

BoxedIter

Boxed iterator type to simplify type annotations