hyperon_atom

Enum Atom

pub enum Atom {
    Symbol(SymbolAtom),
    Expression(ExpressionAtom),
    Variable(VariableAtom),
    Grounded(Box<dyn GroundedAtom>),
}

Atoms are main components of the atomspace. There are four meta-types of atoms: symbol, expression, variable and grounded.

Variants

Symbol(SymbolAtom)

Symbol represents some idea or concept. Two symbols having the same name are considered equal and representing the same concept. Name of the symbol can be arbitrary string. Use Atom::sym to construct new symbol.

Expression(ExpressionAtom)

An expression which may encapsulate other atoms including other expressions. Use Atom::expr to construct new expression.

Variable(VariableAtom)

Variable is used to create patterns. Such pattern can be matched with other atom to assign some specific binding to the variable. Use Atom::var to construct new variable.

Grounded(Box<dyn GroundedAtom>)

Grounded atom represents sub-symbolic data in the atomspace. It may contain any binary object, for example operation, collection or value. Grounded value type creator can define custom type, execution and matching logic for the value (see Grounded). Use Atom::value and Atom::gnd to construct new grounded atom.

Implementations

impl Atom

pub fn iter(&self) -> AtomIter<'_>

Return iterator through all sub-atoms of the Atom.

pub fn iter_mut(&mut self) -> AtomIterMut<'_>

Return mutable iterator through all sub-atoms of the Atom.

impl Atom

pub fn sym<T: Into<UniqueString>>(name: T) -> Self

Constructs new symbol atom with given name.

Examples

use hyperon_atom::Atom;

let a = Atom::sym("A");
let aa = Atom::sym("A");
let b = Atom::sym("B");

assert_eq!(a.to_string(), "A");
assert_eq!(a, aa);
assert_ne!(a, b);

pub fn expr<T: Into<CowArray<Atom>>>(children: T) -> Self

Constructs expression out of array of children.

Examples

use hyperon_atom::Atom;

let expr = Atom::expr([Atom::sym("a"), Atom::sym("b")]);
let same_expr = Atom::expr([Atom::sym("a"), Atom::sym("b")]);
let other_expr = Atom::expr([Atom::sym("+"), Atom::var("x"),
    Atom::expr([Atom::sym("*"), Atom::value(5), Atom::value(8)])]);

assert_eq!(expr.to_string(), "(a b)");
assert_eq!(other_expr.to_string(), "(+ $x (* 5 8))");
assert_eq!(expr, same_expr);
assert_ne!(expr, other_expr);

pub fn var<T: Into<UniqueString>>(name: T) -> Self

Constructs variable out of name.

Examples

use hyperon_atom::Atom;

let a = Atom::var("a");
let aa = Atom::var("a");
let b = Atom::var("b");

assert_eq!(a.to_string(), "$a");
assert_eq!(a, aa);
assert_ne!(a, b);

pub fn gnd<T: CustomGroundedType>(gnd: T) -> Atom

Constructs grounded atom with customized behaviour. See Grounded for examples.

pub fn value<T: AutoGroundedType>(value: T) -> Atom

Constructs grounded atom from Rust value automatically.

Examples

use hyperon_atom::Atom;

let i = Atom::value(1);
let j = Atom::value(1);
let x = Atom::value("b");

assert_eq!(i.to_string(), "1");
assert_eq!(x.to_string(), "\"b\"");
assert_eq!(i, j);
assert_ne!(i, x);

pub fn as_gnd<T: 'static>(&self) -> Option<&T>

Returns reference to the wrapped Rust value of type T if atom is grounded. T should be the exactly the type of the value inside atom.

Examples

use hyperon_atom::Atom;

let x = Atom::value(1u32);

assert_eq!(x.to_string(), "1");
assert_eq!(x.as_gnd::<u32>(), Some(&1u32));
assert_eq!(x.as_gnd::<String>(), None);

pub fn as_gnd_mut<T: 'static>(&mut self) -> Option<&mut T>

Returns mutable reference to the wrapped Rust value of type T if atom is grounded. T should be the exactly the type of the value inside atom.

Examples

use hyperon_atom::Atom;

let mut x = Atom::value(123u32);
assert_eq!(x.to_string(), "123");

*(x.as_gnd_mut::<u32>().unwrap()) = 321u32;
assert_eq!(x.to_string(), "321");