hyperon_atom

Macro bind

macro_rules! bind {
    ($($k:ident: $v:expr),*) => { ... };
}

Constructs new instance of the Bindings with predefined content. Macros takes variable/value pairs as arguments. If value is a single variable then the pair means variable equality. Otherwise pair means assigning value. May be ineffective, should be used mainly in unit tests.

Examples

use hyperon_atom::*;

let bindings = bind!{ a: expr!("A"), b: expr!("foo" "B"), c: expr!(a) };

assert_eq!(bindings.resolve(&VariableAtom::new("a")), Some(expr!("A")));
assert_eq!(bindings.resolve(&VariableAtom::new("b")), Some(expr!("foo" "B")));
assert_eq!(bindings.resolve(&VariableAtom::new("c")), Some(expr!("A")));