hyperon_atom

Macro bind_set

macro_rules! bind_set {
    [] => { ... };
    [{$($b:tt)*}] => { ... };
    [$b:expr] => { ... };
    [$b:expr, $($b_rest:tt)*] => { ... };
}

Constructs a new BindingsSet with predefined content. Macros takes variable/value pairs as arguments. If value is a single variable then the pair means variable equality. Otherwise pair means assigning value. May be ineffective, should be used mainly in unit tests.

Examples

use hyperon_atom::*;

// Compose a BindingsSet with explicitly defined Bindings
let set = bind_set![bind!{a: expr!("A")}, bind!{a: expr!("APrime")}];

assert_eq!(set.len(), 2);
assert_eq!(set[0].resolve(&VariableAtom::new("a")), Some(expr!("A")));
assert_eq!(set[1].resolve(&VariableAtom::new("a")), Some(expr!("APrime")));