hyperon_atom

Module matcher

Module contains functions to match atoms and work with variable bindings.

Structs

Bindings

Represents variable bindings. Keeps two kinds of relations inside: variables equalities and variable value assignments. For example this structure is able to precisely represent result of matching atoms like ($a A C) and ($x $x $y). The result is { $a = $x = A, $y = C }. Bindings contains variables from both sides of the match.

BindingsIter

Iterator over (&VariableAtom, Atom) pairs in Bindings. Each pair contains reference to a VariableAtom and instance of Atom which contains resolved value of the variable. See Bindings::resolve.

BindingsSet

Represents a set of Bindings instances resulting from an operation where multiple matches are possible.

Traits

VariableSet

Abstraction of the variable set. It is used to allow passing both HashSet<&VariableAtom> and HashSet<VariableAtom> to the Bindings::narrow_vars method.

Functions

apply_bindings_to_atom_move

Applies bindings to atom and return it (see apply_bindings_to_atom_mut).

apply_bindings_to_atom_mut

Applies bindings to atom. Function replaces all variables in atom by corresponding bindings.

atoms_are_equivalent

Checks if atoms are equal up to variables replacement.

match_atoms

Matches two atoms and returns an iterator over results. Atoms are treated symmetrically.

Type Aliases

MatchResultIter

Iterator over atom matching results. Each result is an instance of Bindings.