hyperon_atom::matcher

Struct Bindings

pub struct Bindings { /* private fields */ }

Represents variable bindings. Keeps two kinds of relations inside: variables equalities and variable value assignments. For example this structure is able to precisely represent result of matching atoms like ($a A C) and ($x $x $y). The result is { $a = $x = A, $y = C }. Bindings contains variables from both sides of the match.

Implementations

impl Bindings

pub fn new() -> Self

Constructs new empty instance of Bindings.

pub fn len(&self) -> usize

pub fn is_empty(&self) -> bool

Returns true if bindings doesnโt contain any variable.

pub fn resolve(&self, var: &VariableAtom) -> Option<Atom>

Returns value of the variable with all sub-variables resolved using the same binding. Returns None if variable doesnโt have a value assigned or cannot be resolved because of the variable loop.

Examples

use hyperon_atom::*;

let norm_bind = bind!{ a: expr!(("foo" b)), b: expr!("bar") };
let loop_bind = bind!{ a: expr!(("foo" b)), b: expr!(("bar" a)) };
let none_bind = bind!{ a: expr!(("foo" b)) };

assert_eq!(norm_bind.resolve(&VariableAtom::new("a")), Some(expr!(("foo" "bar"))));
assert_eq!(loop_bind.resolve(&VariableAtom::new("a")), None);
assert_eq!(none_bind.resolve(&VariableAtom::new("b")), None);

pub fn add_var_equality( self, a: &VariableAtom, b: &VariableAtom, ) -> Result<Bindings, &'static str>

Asserts equality between two VariableAtoms. If the existing bindings for a and b are incompatible then this method will return an Err. This method will also return an Err if the operation causes split bindings. In situations where split bindings are a possibility, call BindingsSet::add_var_equality instead.

pub fn add_var_binding<T1, T2>( self, var: T1, value: T2, ) -> Result<Bindings, &'static str>
where T1: RefOrMove<VariableAtom>, T2: RefOrMove<Atom>,

Tries to insert value as a binding for the var. If self already has binding for the var and it is not matchable with the value then function returns Err. Otherwise it returns updated Bindings.

Examples

use hyperon_atom::*;
use hyperon_atom::matcher::Bindings;

let a = VariableAtom::new("a");
let b = VariableAtom::new("b");
let c = VariableAtom::new("c");
let mut binds = bind!{ a: expr!("A"), b: expr!("B") };

// Re-asserting an existing binding is ok
binds = binds.add_var_binding(&a, &expr!("A"))?;

// Asserting a conflicting binding is an error
assert!(binds.clone().add_var_binding(&b, &expr!("C")).is_err());

// Creating a new binding is ok
binds = binds.add_var_binding(&c, &expr!("C"))?;

assert_eq!(binds.resolve(&a), Some(expr!("A")));
assert_eq!(binds.resolve(&b), Some(expr!("B")));
assert_eq!(binds.resolve(&c), Some(expr!("C")));

pub fn merge(self, other: &Bindings) -> BindingsSet

Merges b bindings into self if they are compatible. May return a BindingsSet containing multiple Bindings if appropriate. If no compatible bindings can be merged, BindingsSet::empty() will be returned.

Examples

use hyperon_atom::*;
use hyperon_atom::matcher::{Bindings, BindingsSet};

let mut binds = bind!{ a: expr!("A") };
let mut comp = bind!{ b: expr!("B") };
let mut incomp = bind!{ a: expr!("B") };

assert_eq!(binds.clone().merge(&comp), BindingsSet::from(bind!{ a: expr!("A"), b: expr!("B") }));
assert_eq!(binds.merge(&incomp), BindingsSet::empty());

pub fn narrow_vars<T: VariableSet>(&self, vars: &T) -> Bindings

Get narrow bindings which contains only passed set of variables and their dependencies.

Examples

use hyperon_atom::*;
use std::collections::HashSet;

let bindings = bind!{ leftA: expr!("A"), leftA: expr!(rightB),
    leftC: expr!("C"), leftD: expr!(rightE), rightF: expr!("F") };
let right = bindings.narrow_vars(&HashSet::from([&VariableAtom::new("rightB"),
    &VariableAtom::new("rightE"), &VariableAtom::new("rightF")]));

assert_eq!(right, bind!{ rightB: expr!("A"), rightF: expr!("F"), rightE: expr!(rightE) });

pub fn convert_var_equalities_to_bindings( self, preferred_vars: &HashSet<VariableAtom>, ) -> Self

Remove variable equalities from the Bindings and represent them as a variable values. Set of preferred variables is used to select the top level variable name. For example, variable bindings { $a = $b <- A } is converted into { $a <- A, $b <- $a } when preferred vars set is { $a }, and into { $b <- A, $a <- $b } when preferred vars set is { $b }.

pub fn has_loops(&self) -> bool

pub fn iter(&self) -> BindingsIter<'_>

Returns iterator of (&VariableAtom, Atom) pairs to represent Bindings in C API. Each pair contains reference to a VariableAtom and instance of Atom which contains resolved value of the variable. See Bindings::resolve.

Examples

use hyperon_common::assert_eq_no_order;
use hyperon_atom::*;

let bindings = bind!{ leftA: expr!("A"), leftA: expr!(rightB),
    leftC: expr!("C"), leftD: expr!(rightE), rightF: expr!("F") };
let pairs: Vec<(&VariableAtom, Atom)> = bindings.iter().collect();

assert_eq_no_order!(pairs, vec![
    (&VariableAtom::new("leftA"), expr!("A")),
    (&VariableAtom::new("rightB"), expr!("A")),
    (&VariableAtom::new("leftC"), expr!("C")),
    (&VariableAtom::new("leftD"), expr!(leftD)),
    (&VariableAtom::new("rightE"), expr!(leftD)),
    (&VariableAtom::new("rightF"), expr!("F")),
]);

pub fn vars(&self) -> impl Iterator<Item = &VariableAtom>

An iterator visiting all variables in arbitrary order.

pub fn rename_vars<F>(self, rename: F) -> Self
where F: FnMut(VariableAtom) -> VariableAtom,

Rename variables inside bindings using rename.

pub fn apply_and_retain<F>(&mut self, atom: &mut Atom, f: F)
where F: Fn(&VariableAtom) -> bool,