hyperon_atom::matcher

Trait VariableSet

pub trait VariableSet: Debug {
    type Iter<'a>: Iterator<Item = &'a VariableAtom>
       where Self: 'a;

    // Required methods
    fn contains(&self, var: &VariableAtom) -> bool;
    fn iter(&self) -> Self::Iter<'_>;
}

Abstraction of the variable set. It is used to allow passing both HashSet<&VariableAtom> and HashSet<VariableAtom> to the Bindings::narrow_vars method.

Required Associated Types

type Iter<'a>: Iterator<Item = &'a VariableAtom> where Self: 'a

Required Methods

fn contains(&self, var: &VariableAtom) -> bool

Returns true if var is a part of the set.

fn iter(&self) -> Self::Iter<'_>

Iterate trough a list of variables in the set.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.