hyperon_atom

Trait ConvertingSerializer

pub trait ConvertingSerializer<T>: Serializer + Default {
    // Required methods
    fn check_type(gnd: &dyn GroundedAtom) -> bool;
    fn into_type(self) -> Option<T>;

    // Provided method
    fn convert(gnd: &dyn GroundedAtom) -> Result<T, &'static str>
       where T: 'static + Clone { ... }
}

Converts serializable grounded atom into native Rust type T. Returning error from this serializer is not expected as usually it should just write data into a memory buffer for conversion. Default is required to instantiate it in ConvertingSerializer::convert method.

Required Methods

fn check_type(gnd: &dyn GroundedAtom) -> bool

fn into_type(self) -> Option<T>

Provided Methods

fn convert(gnd: &dyn GroundedAtom) -> Result<T, &'static str>
where T: 'static + Clone,

Converts atom into Rust value using Self::default() instance. First it checks whether the grounded atom is already an instance of T and clones value if it the case. Otherwise it uses Self to convert into T via serialization.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.