hyperon_atom

Trait CustomExecute

pub trait CustomExecute {
    // Provided methods
    fn execute(&self, _args: &[Atom]) -> Result<Vec<Atom>, ExecError> { ... }
    fn execute_bindings(
        &self,
        args: &[Atom],
    ) -> Result<BoxedIter<'static, (Atom, Option<Bindings>)>, ExecError> { ... }
}

Trait for implementing custom execution logic. Using this trait one can represent a grounded function as an atom. In order to make it work one should also implement Grounded::as_execute method.

Examples

use hyperon_atom::*;
use std::fmt::{Display, Formatter};
use std::iter::once;

#[derive(Debug, PartialEq, Clone)]
struct MyGrounded {}

impl Grounded for MyGrounded {
    fn type_(&self) -> Atom {
        rust_type_atom::<MyGrounded>()
    }

    fn as_execute(&self) -> Option<&dyn CustomExecute> {
        Some(self)
    }
}

impl CustomExecute for MyGrounded {
    fn execute(&self, args: &[Atom]) -> Result<Vec<Atom>, ExecError> {
        Ok(vec![sym!("result")])
    }
}

impl Display for MyGrounded {
    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
        write!(f, "MyGrounded")
    }
}

let atom = Atom::gnd(MyGrounded{});
let gnd = if let Atom::Grounded(ref gnd) = atom { gnd } else { panic!("Non grounded atom"); };

assert_eq!(gnd.as_grounded().as_execute().unwrap().execute(&mut vec![]), Ok(vec![sym!("result")]));

Provided Methods

fn execute(&self, _args: &[Atom]) -> Result<Vec<Atom>, ExecError>

Executes grounded function on passed args and returns list of results as Vec<Atom> or ExecError. This is a restricted version of the CustomExecute::execute_bindings method which is implemented if returning of variable bindings is not required.

fn execute_bindings( &self, args: &[Atom], ) -> Result<BoxedIter<'static, (Atom, Option<Bindings>)>, ExecError>

Executes grounded function on passed args and returns list of results with associated variable bindings if there are ones. Default implementation calls CustomExecute::execute and returns no bindings.