hyperon_atom

Trait CustomMatch

pub trait CustomMatch {
    // Required method
    fn match_(&self, other: &Atom) -> MatchResultIter;
}

Trait for implementing custom matching logic. In order to make it work one should also implement Grounded::as_match method.

Examples

use hyperon_atom::*;
use hyperon_atom::matcher::{BindingsSet, Bindings, MatchResultIter, match_atoms};
use std::fmt::{Display, Formatter};
use std::iter::once;

#[derive(Debug, PartialEq, Clone)]
struct MyGrounded {}

impl Grounded for MyGrounded {
    fn type_(&self) -> Atom {
        rust_type_atom::<MyGrounded>()
    }

    fn as_match(&self) -> Option<&dyn CustomMatch> {
        Some(self)
    }
}

impl CustomMatch for MyGrounded {
    fn match_(&self, other: &Atom) -> MatchResultIter {
        // Doesn't match with anything even with itself
        Box::new(BindingsSet::empty().into_iter())
    }
}

impl Display for MyGrounded {
    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
        write!(f, "MyGrounded")
    }
}

let atom = Atom::gnd(MyGrounded{});
let same = Atom::gnd(MyGrounded{});

assert_eq!(atom, same);
assert_eq!(match_atoms(&atom, &same).collect::<Vec<Bindings>>(), vec![]);

Required Methods

fn match_(&self, other: &Atom) -> MatchResultIter

Implements custom matching logic of the grounded atom. Gets other atom as input, returns the iterator of the matcher::Bindings for the variables of the other atom. See matcher for detailed explanation.