hyperon_atom

Trait Grounded

pub trait Grounded: Display {
    // Required method
    fn type_(&self) -> Atom;

    // Provided methods
    fn as_execute(&self) -> Option<&dyn CustomExecute> { ... }
    fn as_match(&self) -> Option<&dyn CustomMatch> { ... }
    fn serialize(&self, _serializer: &mut dyn Serializer) -> Result { ... }
}

Trait allows implementing grounded atom with custom behavior. If no custom behavior is needed then simpler way is using Atom::value function for automatic grounding.

Examples

use hyperon_atom::*;
use hyperon_atom::matcher::{Bindings, match_atoms};
use std::fmt::{Display, Formatter};
use std::iter::once;

#[derive(Debug, PartialEq, Clone)]
struct MyGrounded {}

impl Grounded for MyGrounded {
    fn type_(&self) -> Atom {
        rust_type_atom::<MyGrounded>()
    }
}

impl Display for MyGrounded {
    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
        write!(f, "MyGrounded")
    }
}

let atom = Atom::gnd(MyGrounded{});
let other = Atom::gnd(MyGrounded{});
let gnd = if let Atom::Grounded(ref gnd) = atom { gnd } else { panic!("Non grounded atom"); };

println!("{}", gnd.type_());

assert_eq!(atom.to_string(), "MyGrounded");
assert_ne!(atom, Atom::sym("MyGrounded"));
assert!(gnd.as_grounded().as_execute().is_none());
assert_eq!(match_atoms(&atom, &other).collect::<Vec<Bindings>>(), vec![Bindings::new()]);
assert_eq!(atom, other);

Required Methods

fn type_(&self) -> Atom

Returns type of the grounded atom. Should return same type each time it is called. rust_type_atom function can be used to implement default behavior if required.

Provided Methods

fn as_execute(&self) -> Option<&dyn CustomExecute>

Returns reference to the custom execution API implementation. If None is returned then atom is not executable. See CustomExecute for details.

fn as_match(&self) -> Option<&dyn CustomMatch>

Returns reference to the custom matching API implementation. If None is returned then atom is matched by equality. See CustomMatch for details.

fn serialize(&self, _serializer: &mut dyn Serializer) -> Result

Implements serialization logic of the grounded atom. The logic is implemented in terms of the Rust native types. See serial for details.