hyperon_atom

Macro expr

macro_rules! expr {
    () => { ... };
    ($x:ident) => { ... };
    ($x:literal) => { ... };
    ({$x:expr}) => { ... };
    (($($x:tt)*)) => { ... };
    ($($x:tt)*) => { ... };
}

Constructs new Atom using symplified syntax for expressions. Macros has a perfomance penalty because it creates and uses an additional wrapper for grounded atoms. It is intended for using mainly in the unit tests. Use string literals for symbols, identifiers for variables, braces for grounded symbols and S-expressions syntax for expressions.

Examples

#[macro_use]
use hyperon_atom::{expr, Atom, ExecError};
use hyperon_atom::gnd::GroundedFunctionAtom;

let mul = GroundedFunctionAtom::new("*".into(),
    expr!("->" "Number" "Number" "Number"),
    |args: &[Atom]| -> Result<Vec<Atom>, ExecError> { Err(ExecError::NoReduce) });

let sym = expr!("A");
let var = expr!(x);
let gnd = expr!({42});
let expr = expr!("=" ("*2" n) ({mul} n {2}));

assert_eq!(sym.to_string(), "A");
assert_eq!(var.to_string(), "$x");
assert_eq!(gnd.to_string(), "42");
assert_eq!(expr.to_string(), "(= (*2 $n) (* $n 2))");